refactor dashboard code to single components
